Enable Sharing by Email Address
Update the CLI sharing command to support using an email address for client identification (the passed parameter will be checked to see if it's an email and used to fetch client info for the client ID if so).

Also, add in some minimal email validation during registration (closing out a pending TODO).

Finally, make emails discoverable by default to extend the usability of the CLI as a demo.