Clean up Naming and Channels
Make the constructor for an `EventSource` more precise and use the name "Event Source" to represent what it is we're building (as the fact that it's a websocket is abstracted from the client. Also, update the `EventSource` such that the actual events channel is exposed as a one-way channel (versus bi-directional) via a getter.