More Flexible Plaintext Handling
An oversight in the PDS service allowed setting field values in plaintext meta to non-strings. Though we explicitly document the `plain` field as being a map of String->String, end users might make the same mistake down the road.

This patch allows for a map of String->Any in Go. It will safely handle both Strings and Integers moving forward.