[WASH-1124] add TozIDRealmIDPAccessToken to SDK
It is a jwt for the tozid-realm-idp agent client which is returned in the
final redirect response of the login flow when logging into the account client.
Js-sdk calls this the "agent token".

An auth pattern exists in keycloak to use it to maintain a user's session
across API calls. It will be needed for automated testing of WebAuthn MFA
and for supporting FIDO2 devices in TODA.